Added RGB led support for different colors
board/pin_mux.c
To accomplish this it was required to enable the clock for
PortB and E, and configure the appropriate pins.
(Red is PTB22, Green is PTE26 and Blue is PTB21).

source/gpio_input_interrupt.c
It was added to the application the usage of the three leds
Before the infinite loop the Leds are turn off as an
initial state.

Signed-off-by: Marcela Rosales <marcelarosalesj@gmail.com>